refactor(review): isolate bounded recovery reads
