+++ src/autoskillit/pipeline/context_admission_ledger.py
@@ -23,7 +23,6 @@ from typing import Any, Final, assert_never, cast, get_args
 from autoskillit.core import (
     CONTEXT_ADMISSION_ENCODING_VERSION,
     CONTEXT_ADMISSION_PROTOCOL_VERSION,
-    CONTEXT_ADMISSION_TOP_LEVEL_DISCRIMINATORS,
     AcceptInputEvent,
     ActiveContextAdmissionState,
     AdmissionBatch,
@@ -79,13 +78,18 @@ from autoskillit.core import (
     StageHistoryEvent,
     StartGenerationEvent,
     UninitializedContextAdmissionState,
-    context_admission_envelope_header,
     context_admission_reducer_for_protocol,
     decode_stored_context_admission_envelope,
     encode_stored_context_admission_envelope,
     make_stored_context_admission_envelope,
 )
 
+from ._context_admission_recovery import (
+    _LedgerOpenError,
+    _LedgerReadBudget,
+    _preflight_storage_routes,
+    _read_bounded_rows,
+)
 from ._context_admission_storage import (
     SCHEMA_SQL as _SCHEMA_SQL,
 )
@@ -137,52 +141,8 @@ class _LedgerFaultPoint(StrEnum):
     AFTER_COMMIT = "after_commit"
 
 
-class _LedgerOpenError(RuntimeError):
-    def __init__(
-        self,
-        reason: ContextAdmissionStorageFailureReason,
-        reason_code: str,
-    ) -> None:
-        super().__init__(reason_code)
-        self.reason = reason
-        self.reason_code = reason_code
-
-
 class _LedgerContended(RuntimeError):
     pass
-
-
-class _LedgerReadBudget:
-    __slots__ = ("_bytes", "_reason_code", "_rows")
-
-    def __init__(self, reason_code: str) -> None:
-        self._rows = 0
-        self._bytes = 0
-        self._reason_code = reason_code
-
-    def consume(self, row: tuple[Any, ...]) -> tuple[Any, ...]:
-        self._rows += 1
-        self._bytes += sum(
-            len(value)
-            if isinstance(value, bytes | bytearray | memoryview)
-            else len(value.encode("utf-8"))
-            if isinstance(value, str)
-            else 0
-            for value in row
-        )
-        if self._rows > _MAX_RECOVERY_ROWS or self._bytes > _MAX_RECOVERY_BYTES:
-            raise _LedgerOpenError(
-                ContextAdmissionStorageFailureReason.INTEGRITY,
-                self._reason_code,
-            )
-        return row
-
-
-def _read_bounded_rows(
-    cursor: sqlite3.Cursor,
-    budget: _LedgerReadBudget,
-) -> tuple[tuple[Any, ...], ...]:
-    return tuple(budget.consume(cast(tuple[Any, ...], row)) for row in cursor)
 
 
 class DefaultContextAdmissionLedger:
@@ -817,7 +777,11 @@ class DefaultContextAdmissionLedger:
                     sqlite3.SQLITE_LIMIT_LENGTH,
                     max(1, _MAX_RECOVERY_BYTES),
                 )
-                read_budget = _LedgerReadBudget("recovery-read-limit-exceeded")
+                read_budget = _LedgerReadBudget(
+                    "recovery-read-limit-exceeded",
+                    max_rows=_MAX_RECOVERY_ROWS,
+                    max_bytes=_MAX_RECOVERY_BYTES,
+                )
                 self._validate_integrity(connection)
                 metadata = dict(
                     _read_bounded_rows(
@@ -997,7 +961,11 @@ class DefaultContextAdmissionLedger:
                     sqlite3.SQLITE_LIMIT_LENGTH,
                     max(1, _MAX_RECOVERY_BYTES),
                 )
-                read_budget = _LedgerReadBudget("inspection-read-limit-exceeded")
+                read_budget = _LedgerReadBudget(
+                    "inspection-read-limit-exceeded",
+                    max_rows=_MAX_RECOVERY_ROWS,
+                    max_bytes=_MAX_RECOVERY_BYTES,
+                )
                 row = connection.execute(
                     """
                     SELECT stream_key, state_envelope, latest_journal_sequence,
@@ -1597,57 +1565,6 @@ def _stored_stream_health(
         ),
         reason_code=str(reason_code) if reason_code is not None else None,
     )
-
-
-def _preflight_storage_routes(
-    connection: sqlite3.Connection,
-    read_budget: _LedgerReadBudget,
-) -> None:
-    queries = (
-        "SELECT genesis_envelope FROM streams",
-        "SELECT state_envelope FROM streams",
-        "SELECT event_envelope FROM journal_events",
-        "SELECT decision_envelope FROM journal_events",
-        "SELECT effect_envelope FROM effect_outbox",
-        "SELECT shadow_envelope FROM shadow_decisions",
-    )
-    for query in queries:
-        for (encoded,) in _read_bounded_rows(connection.execute(query), read_budget):
-            encoded_bytes = bytes(encoded)
-            encoding_version, protocol_version, discriminator = _envelope_header(encoded_bytes)
-            if encoding_version != CONTEXT_ADMISSION_ENCODING_VERSION:
-                try:
-                    envelope = decode_stored_context_admission_envelope(encoded_bytes)
-                except ContextAdmissionValidationError as exc:
-                    raise _LedgerOpenError(
-                        ContextAdmissionStorageFailureReason.UNSUPPORTED_ENCODING,
-                        "unsupported-envelope-encoding",
-                    ) from exc
-                encoding_version = envelope.encoding_version
-                protocol_version = envelope.protocol_version
-                discriminator = envelope.type_discriminator
-            if discriminator not in CONTEXT_ADMISSION_TOP_LEVEL_DISCRIMINATORS:
-                raise _LedgerOpenError(
-                    ContextAdmissionStorageFailureReason.UNSUPPORTED_ENCODING,
-                    "unsupported-envelope-discriminator",
-                )
-            try:
-                context_admission_reducer_for_protocol(protocol_version)
-            except ContextAdmissionValidationError as exc:
-                raise _LedgerOpenError(
-                    ContextAdmissionStorageFailureReason.UNSUPPORTED_PROTOCOL,
-                    "unsupported-envelope-protocol",
-                ) from exc
-
-
-def _envelope_header(value: bytes) -> tuple[int, int, str]:
-    try:
-        return context_admission_envelope_header(value)
-    except ContextAdmissionValidationError:
-        raise _LedgerOpenError(
-            ContextAdmissionStorageFailureReason.INTEGRITY,
-            "invalid-envelope-header",
-        ) from None
 
 
 def _recover_stream_projection(
